% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_window.R
\name{Ar1}
\alias{Ar1}
\title{Lag-1 autocorrelation with detrending}
\usage{
Ar1(x, detrend = FALSE, prop_req = 0.99)
}
\arguments{
\item{x}{numeric}

\item{detrend}{TRUE or FALSE, default = FALSE}

\item{prop_req}{numeric value between 0 and 1, default = 0.99}
}
\value{
numeric
}
\description{
Lag-1 autocorrelation with detrending
}
\examples{
y = rnorm(10)
Ar1(y)
Ar1(y, detrend = TRUE)
y[1] = NA
Ar1(y)
Ar1(y, prop_req = 0.5)
}
