% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_window.R
\name{Ar1_sd}
\alias{Ar1_sd}
\title{SD of lag-1 autocorrelation with detrending}
\usage{
Ar1_sd(x, detrend = FALSE, prop_req = 0.99)
}
\arguments{
\item{x}{numeric}

\item{detrend}{TRUE or FALSE, default = FALSE}

\item{prop_req}{numeric value between 0 and 1, default = 0.99}
}
\value{
numeric
}
\description{
SD of lag-1 autocorrelation with detrending
}
\examples{
y <- rnorm(10)
Ar1_sd(y)
Ar1_sd(y, detrend = TRUE)
y[1] <- NA
Ar1_sd(y)
Ar1_sd(y, prop_req = 0.5)
}
