% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bloom_fert_dates}
\alias{bloom_fert_dates}
\title{Dates that experimental fertilization started and ended on, and that blooms started}
\format{
A data frame with 9 rows and 5 variables:
\describe{
  \item{Lake}{character indicating lake; "L" is Paul Lake, "R" is Peter Lake, "T" is Tuesday Lake}
  \item{Year}{year}
  \item{fertStartDOY}{"Day Of Year" or Julian day of year that nutrient additions to a lake started on}
  \item{fertEndDOY}{"Day Of Year" or Julian day of year that nutrient additions to a lake ended on}
  \item{bloomStartDOY}{"Day Of Year" or Julian day of year that bloom threshold was first crossed}
}
}
\source{
\url{Cascade research group; TODO add URL}
}
\usage{
bloom_fert_dates
}
\description{
A dataset containing dates (Year and Day Of Year = DOY) that algal blooms started (crossed lake-specific thresholds) and that fertilizations started and ended.
}
\keyword{datasets}
