% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickest_detection.R
\name{calc_alarm_rates}
\alias{calc_alarm_rates}
\title{Calculate rates of true and false alarms}
\usage{
calc_alarm_rates(qd_alarms, bloom_fert_df = bloom_fert_dates)
}
\arguments{
\item{qd_alarms}{data frame of quickest detection alarms, output from \code{\link{format_qd}}}

\item{bloom_fert_df}{data frame of fertilization start and end dates, see \code{\link{bloom_fert_dates}} for default and formatting}
}
\value{
data frame with true and false positive rates for combinations of variable and statistic
}
\description{
Calculate rates of true and false alarms
}
\examples{
use_vars <- c("Manual_Chl", "BGA_HYLB", "DO_Sat", "pH")
rw_stats <- calc_rolling_stats(ts_data, var_cols = use_vars)
qd_stats <- run_qd(rw_stats, var_cols = use_vars)
qd_alarms <- format_qd(qd_stats, bloom_fert_df = bloom_fert_dates)
calc_alarm_rates(qd_alarms)
}
