% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_stats.R
\name{calc_moransI}
\alias{calc_moransI}
\title{Calculate Moran's I from a data frame}
\usage{
calc_moransI(
  df_sssv,
  lon_col = "longitude",
  lat_col = "latitude",
  meas_col = "Measurement"
)
}
\arguments{
\item{df_sssv}{data frame with spatial measurements of a single sample and single variable}

\item{lon_col}{character string, column name for longitude column}

\item{lat_col}{character string, column name for latitude column}

\item{meas_col}{character string, column name for measurement column}
}
\value{
numeric, the Moran's I (spatial autocorrelation)
}
\description{
Computes spatial autocorrelation for a single sample and single variable (sssv) from a data frame with three columns (longitude, latitude, and measurement). Uses inverse of distance between points in Moran's I calculation.
}
\examples{
example_data <- data.frame(
  lon = runif(n = 10, min = -90, max = -85),
  lat = runif(n = 10, min = 45, max = 46),
  meas = rnorm(10)
)
calc_moransI(example_data, lon_col = "lon", lat_col = "lat", meas_col = "meas")

peter_lake_data <- flame_data \%>\%
  dplyr::filter(Lake == "R" & Date == as.Date("2019-07-22")) \%>\%
  dplyr::select(latitude, longitude, BGApc_ugL_tau)
calc_moransI(peter_lake_data, meas_col = "BGApc_ugL_tau")
}
