% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickest_detection.R
\name{f_corr_coef}
\alias{f_corr_coef}
\title{Exact density for Pearson's correlation coefficient}
\usage{
f_corr_coef(r, rho, N)
}
\arguments{
\item{r}{numeric, sample correlation}

\item{rho}{numeric, "true" correlation coefficient}

\item{N}{numeric, sample size}
}
\value{
numeric, the probability density at *r*
}
\description{
Exact density for Pearson's correlation coefficient
}
\examples{
f_corr_coef(r = 0.5, rho = 0.7, N = 10)
}
