% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flame_data}
\alias{flame_data}
\title{Spatial data of lake bloom variables}
\format{
A data frame with 176291 rows and 10 variables:
\describe{
  \item{Lake}{character indicating lake; "L" is Paul Lake, "R" is Peter Lake, "T" is Tuesday Lake}
  \item{Year}{year}
  \item{DOY}{"Day Of Year" or Julian day of year that sample was collected on, integer from 1 to 366}
  \item{Date}{Date that sample was collected on, Date (YYYY-MM-DD) format}
  \item{date_time}{precise data-time that sample was collected at, POSIXct format in UTC time zone}
  \item{latitude}{latitude of sample}
  \item{longitude}{longitude of sample}
  \item{BGApc_ugL_tau}{blue green algae fluorescence measured by YSI EXO3 sonde, in units ug/L of phycocyanin}
  \item{ODO_percent_tau}{dissolved oxygen, as percent of saturation in equilibrium with atmosphere}
  \item{pH_tau}{pH measured by sonde}
}
}
\source{
\url{Cascade research group; TODO add URL}
}
\usage{
flame_data
}
\description{
A dataset containing spatial data of bloom variables collected using FLAMe spatial sampling system (TODO: add FLAMe url). Measurements are 'tau' corrected, corrected for the hydrology residence time of the system and each sensor's response time.
}
\keyword{datasets}
