% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_window.R
\name{Kurtosis}
\alias{Kurtosis}
\title{Calculate the kurtosis of values with optional (linear) detrending, and perform check that there's enough data}
\usage{
Kurtosis(x, detrend = FALSE, prop_req = 0.99)
}
\arguments{
\item{x}{numeric}

\item{detrend}{TRUE or FALSE, default = FALSE}

\item{prop_req}{numeric value between 0 and 1, default = 0.99}
}
\value{
numeric
}
\description{
Calculate the kurtosis of values with optional (linear) detrending, and perform check that there's enough data
}
\examples{
y <- rnorm(10)
Kurtosis(y)
Kurtosis(y, detrend = TRUE)
y[1] <- NA
Kurtosis(y)
Kurtosis(y, prop_req = 0.5)
}
