% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_window.R
\name{Mean}
\alias{Mean}
\title{Mean with detrending}
\usage{
Mean(X, detrend = FALSE, prop_req = 0.99)
}
\arguments{
\item{X}{numeric}

\item{detrend}{TRUE or FALSE, default = FALSE}

\item{prop_req}{numeric value between 0 and 1, default = 0.99}
}
\value{
numeric
}
\description{
Calculate the mean of values with optional (linear) detrending, and perform check that there's enough data
}
\examples{
y = rnorm(10)
Mean(y)
Mean(y, detrend = TRUE)
y[1] = NA
Mean(y)
Mean(y, prop_req = 0.5)
}
