% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peter_paul_bounds}
\alias{peter_paul_bounds}
\title{Boundaries for study lakes}
\format{
An sf data frame with 2 rows and 15 variables:
\describe{
  \item{AREA}{numeric, lake area in square meters}
  \item{PERIMETER}{numeric, lake perimenter in meters meters}
  \item{LAKE_TYPE}{TODO: document}
  \item{NAME}{TODO: document}
  \item{UNIQUE_ID}{TODO: document}
  \item{COUNTY}{TODO: document}
  \item{NOTE24}{TODO: document}
  \item{NEW_KEY}{TODO: document}
  \item{HECTARES}{TODO: document}
  \item{ACRES_GIS}{TODO: document}
  \item{FMU}{TODO: document}
  \item{Lake}{TODO: document}
  \item{FID_1}{TODO: document}
  \item{Lake_Name}{TODO: document}
  \item{geometry}{TODO: document}

}
}
\source{
\url{TODO add where this came from}
}
\usage{
peter_paul_bounds
}
\description{
An sf data.frame with boundaries for study lakes, Peter and Paul lakes.
}
\keyword{datasets}
