% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_manuscript.R
\name{plot_fig1}
\alias{plot_fig1}
\title{Plot state variables}
\usage{
plot_fig1(
  time_series = ts_data,
  bloom_fert_df = bloom_fert_dates,
  var_rename_vec = c(`Chl-a (ug/L)` = "Manual_Chl", `BGA (cells/mL)` = "BGA_HYLB",
    `D.O. sat. (\%)` = "DO_Sat", pH = "pH"),
  lakes_rename_vec = c(Reference = "L", Experimental = "R"),
  lake_colors = c(Experimental = "firebrick3", Reference = "royalblue3"),
  legend_location = c(0.13, 0.9)
)
}
\arguments{
\item{time_series}{data frame containing time series of state variables, , see \code{\link{ts_data}} for default and formatting}

\item{bloom_fert_df}{data frame of fertilization start and end and bloom start dates, see \code{\link{bloom_fert_dates}} for default and formatting}

\item{var_rename_vec}{named vector, used to change variable names displayed in plot, format is c("new name" = "original name")}

\item{lakes_rename_vec}{named vector, used to change lake names displayed in plot, format is c("new name" = "original name")}

\item{lake_colors}{named vector, used to specify colors for each lake, format is c("Lake" = "color")}

\item{legend_location}{numerical vector of length 2, has c(x, y) position for placing legend on plot}
}
\value{
ggplot object plotting the time series
}
\description{
Plot state variables
}
\examples{
# all data:
plot_fig1()

# just latest experiment
plot_fig1(
  time_series = ts_data \%>\% dplyr::filter(Year >= 2018),
  bloom_fert_df = bloom_fert_dates \%>\% dplyr::filter(Lake == "R" & Year >= 2018)
)
}
