% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_manuscript.R
\name{plot_fig2}
\alias{plot_fig2}
\title{Plot example maps of spatial data}
\usage{
plot_fig2(
  spatial_data = flame_data,
  lake_boundaries = peter_paul_bounds,
  samples_to_plot = data.frame(Year = 2019, DOY = c(165, 210, 228), stringsAsFactors =
    FALSE),
  var_cols = c("BGApc_ugL_tau", "ODO_percent_tau", "pH_tau")
)
}
\arguments{
\item{spatial_data}{data frame containing spatial locations and measurements of one or more variables; see \code{\link{flame_data}} for default and formatting}

\item{lake_boundaries}{sf data frame containing boundaries for lakes; see \code{\link{peter_paul_bounds}} and str(peter_paul_bounds) for default and formatting}

\item{samples_to_plot}{a data frame specifying which info from \code{spatial_data} should be used for plotting}

\item{var_cols}{character vector, column names that hold measurements of different variables}
}
\value{
a ggplot object with the specified maps of spatial data
}
\description{
Plot example maps of spatial data
}
\examples{
plot_fig2(
  spatial_data = flame_data,
  samples_to_plot = data.frame(Year = 2019, DOY = c(165, 210, 228), stringsAsFactors = FALSE),
  var_cols = c("BGApc_ugL_tau")
)
}
