% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_manuscript.R
\name{plot_fig3}
\alias{plot_fig3}
\title{Plot rolling window stats and quickest detection alarms for all variables}
\usage{
plot_fig3(
  rolling_window_stats,
  qd_alarms,
  bloom_dates,
  rename_vec = c(`Chl-a` = "Manual_Chl", BGA = "BGA_HYLB", `D.O. sat.` = "DO_Sat", pH =
    "pH")
)
}
\arguments{
\item{rolling_window_stats}{data frame, output from call to}

\item{qd_alarms}{data frame, output from call to}

\item{bloom_dates}{data frame, formatted as}
}
\value{
ggplot object showing rolling window statistics and quickest detection alarms
}
\description{
Plot rolling window stats and quickest detection alarms for all variables
}
\examples{
use_vars = c("Manual_Chl", "BGA_HYLB", "DO_Sat", "pH")
rw_stats = calc_rolling_stats(ts_data, var_cols=use_vars)
qd_s = run_qd(rw_stats, var_cols=use_vars)
qd_a = format_qd(qd_s, bloom_fert_df = bloom_fert_dates)
plot_fig3(rolling_window_stats = rw_stats, qd_alarms = qd_a, bloom_dates = bloom_fert_dates)
}
