% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_manuscript.R
\name{plot_fig4}
\alias{plot_fig4}
\title{Plot spatial statistics time series}
\usage{
plot_fig4(
  spatial_stats,
  bloom_fert_df = bloom_fert_dates,
  var_rename_vec = c(`BGA (ug/L)` = "BGApc_ugL_tau", `DO sat.` = "ODO_percent_tau", pH
    = "pH_tau")
)
}
\arguments{
\item{spatial_stats}{data frame, output from call to \code{\link{calc_spatial_stats}}}

\item{bloom_fert_df}{data frame of fertilization start and end and bloom start dates, see \code{\link{bloom_fert_dates}} for default and formatting}

\item{var_rename_vec}{named vector, used to change variable names displayed in plot, format is c("new name" = "original name")}
}
\value{
a ggplot object showing the change in spatial statistics over the experiment
}
\description{
Plot spatial statistics time series
}
\examples{
spat_stats <- calc_spatial_stats(
  spatial_data = flame_data
)
plot_fig4(spat_stats)
}
