% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{plot_FLAMe_maps}
\alias{plot_FLAMe_maps}
\title{Plot spatial maps from FLAMe system}
\usage{
plot_FLAMe_maps()
}
\value{

}
\description{
Run this function (with no arguments) to launch a shiny application that allows for visualizing spatial measurements for user-chosen dates and variables.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  plot_FLAMe_maps()
}
}
