% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickest_detection.R
\name{plot_qd}
\alias{plot_qd}
\title{Plot rolling window statistics and qd alarms}
\usage{
plot_qd(
  rolling_window_stats,
  qd_alarms,
  var_col,
  stats = c("SD", "Ar1"),
  title = "R.W. width = 21",
  widths = c(21),
  plot_bloom_lines = FALSE,
  bloom_dates
)
}
\arguments{
\item{rolling_window_stats}{data frame of rolling window statistics that is output from \code{\link{calc_rolling_stats}}}

\item{qd_alarms}{data frame of quickest detection alarms, output from \code{\link{format_qd}}}

\item{var_col}{character vector, variable names that are columns of *data* containing time series to calculate rolling window stats on}

\item{stats}{statistics to plot, default and options are c("SD", "Ar1")}

\item{title}{title to put on the plot}

\item{widths}{numeric vector, rolling window width(s) to be used, default is c(21)}

\item{plot_bloom_lines}{TRUE or FALSE (default), should lines for the bloom dates be plotted?}

\item{bloom_dates}{if plot_bloom_lines = TRUE, this data frame is used to determine where}
}
\value{
a ggplot object
}
\description{
Plot rolling window statistics and qd alarms
}
\examples{
use_vars <- c("Manual_Chl", "BGA_HYLB", "DO_Sat", "pH")
rw_stats <- calc_rolling_stats(ts_data, var_cols = use_vars)
qd_stats <- run_qd(rw_stats, var_cols = use_vars)
qd_alarms <- format_qd(qd_stats, bloom_fert_df = bloom_fert_dates)
plot_qd(rw_stats, qd_alarms, use_vars[1],
  stats = c("SD", "Ar1"),
  plot_bloom_lines = TRUE, bloom_dates = bloom_fert_dates
)
}
