% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickest_detection.R
\name{run_qd}
\alias{run_qd}
\title{Quickest Detection calculation}
\usage{
run_qd(
  rolling_window_stats,
  var_cols,
  widths = c(21),
  stats_to_qd = c("SD", "Ar1"),
  A_adj = 1e+07,
  exp_lakes = c("R", "T"),
  ref_lake = "L",
  ar1_alarm_rho = 0.95
)
}
\arguments{
\item{rolling_window_stats}{data frame of rolling window statistics that is output from \code{\link{calc_rolling_stats}}}

\item{var_cols}{character vector, variable names that are columns of *data* containing time series to calculate quickest detection stats on}

\item{widths}{numeric vector, rolling window width(s) to be used, default is c(21)}

\item{stats_to_qd}{character vector, rolling window statistics to run quickest detection method on, defaults to all available: c("sd", "Ar1")}

\item{A_adj}{numeric, the threshold in the S-R statistic at which an alarm is triggered}

\item{exp_lakes}{character vector, values in *Lake* column of *rolling_window_stats* that correspond to experimental lakes}

\item{ref_lake}{character vector, values in *Lake* column of *rolling_window_stats* that correspond to the reference lake}

\item{ar1_alarm_rho}{numeric value between 0 and 1 (exclusive), 'true' value of lag-1 autocorrelation in alarm state, default is 0.95}
}
\value{
a data frame of values from the quickest detection method; TODO add more details on columns
}
\description{
Quickest Detection calculation
}
\examples{
rw_stats = calc_rolling_stats(ts_data, var_cols=c("DO_Sat"))
qd_stats = run_qd(rw_stats, var_cols=c("DO_Sat"))
}
