% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_window.R
\name{SD}
\alias{SD}
\title{SD with detrending}
\usage{
SD(x, detrend = FALSE, prop_req = 0.99)
}
\arguments{
\item{x}{numeric}

\item{detrend}{TRUE or FALSE, default = FALSE}

\item{prop_req}{numeric value between 0 and 1, default = 0.99}
}
\value{
numeric
}
\description{
Calculate the standard deviation of values with optional (linear) detrending, and perform check that there's enough data
}
\examples{
y <- rnorm(10)
SD(y)
SD(y, detrend = TRUE)
y[1] <- NA
SD(y)
SD(y, prop_req = 0.5)
}
