% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_window.R
\name{SD_sd}
\alias{SD_sd}
\title{SD of SD autocorrelation with detrending}
\usage{
SD_sd(x, detrend = FALSE, prop_req = 0.99)
}
\arguments{
\item{x}{numeric}

\item{detrend}{TRUE or FALSE, default = FALSE}

\item{prop_req}{numeric value between 0 and 1, default = 0.99}
}
\value{
numeric
}
\description{
SD of SD autocorrelation with detrending
}
\examples{
y = rnorm(10)
SD_sd(y)
SD_sd(y, detrend = TRUE)
y[1] = NA
SD_sd(y)
SD_sd(y, prop_req = 0.5)
}
