% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_window.R
\name{Skewness}
\alias{Skewness}
\title{Calculate the skewness of values with optional (linear) detrending, and perform check that there's enough data}
\usage{
Skewness(x, detrend = FALSE, prop_req = 0.99)
}
\arguments{
\item{x}{numeric}

\item{detrend}{TRUE or FALSE, default = FALSE}

\item{prop_req}{numeric value between 0 and 1, default = 0.99}
}
\value{
numeric
}
\description{
Calculate the skewness of values with optional (linear) detrending, and perform check that there's enough data
}
\examples{
y <- rnorm(10)
Skewness(y)
Skewness(y, detrend = TRUE)
y[1] <- NA
Skewness(y)
Skewness(y, prop_req = 0.5)
}
