% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_data}
\alias{ts_data}
\title{Daily time series data of lake bloom variables}
\format{
A data frame with 1957 rows and 7 variables:
\describe{
  \item{Lake}{character indicating lake; "L" is Paul Lake, "R" is Peter Lake, "T" is Tuesday Lake}
  \item{Year}{year}
  \item{DOY}{"Day Of Year" or Julian day of year, integer from 1 to 366}
  \item{BGA_HYLB}{blue green algae fluorescence measured by Hydrolab DS5X sonde, in units cells/mL}
  \item{Manual_Chl}{chlorophyll concentration measured by manual extraction and lab fluorometry, units ug/L}
  \item{DO_Sat}{dissolved oxygen, as percent of saturation in equilibrium with atmosphere}
  \item{pH}{pH measured by sonde}
}
}
\source{
\url{Cascade research group; TODO add URL}
}
\usage{
ts_data
}
\description{
A dataset containing daily data of bloom variables
}
\keyword{datasets}
