% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_manuscript.R
\name{plot_fig5}
\alias{plot_fig5}
\title{Plot true and false positive quickest detection alarm rates}
\usage{
plot_fig5(qd_alarm_rates)
}
\arguments{
\item{qd_alarm_rates}{data frame, output from call to}
}
\value{

}
\description{
Plot true and false positive quickest detection alarm rates
}
\examples{
# need to add
}
